requirements.txt
python-telegram-bot==21.10
mysql-connector-python==9.2.0
paramiko==3.5.1
python-dotenv==1.1.0
requests==2.32.3
beautifulsoup4==4.13.3
